+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [fix: fix count circuit breaker in gateway & return 404 when context api does not match.](https://github.com/Tencent/spring-cloud-tencent/pull/1529)
 - [fix:fix watch tsf config, fix bean refresh with RefreshScope and ConfigurationProperties.](https://github.com/Tencent/spring-cloud-tencent/pull/1530)
 - [docs:simplify GitHub Actions.](https://github.com/Tencent/spring-cloud-tencent/pull/1531)
+- [feat: support config event.](https://github.com/Tencent/spring-cloud-tencent/pull/1534)